fix: 修复菜单显示错乱和类型转换警告
- 修复从摄像头页面返回主菜单后滚动渲染显示错乱的问题
- 在快捷跳转到Cargo时添加need_refresh标志，确保屏幕正确刷新
- 添加otsu_get_threshold返回值的显式uint8类型转换，消除编译警告

+++ code/menu_config.c
@@ -494,7 +494,7 @@ void camera_display_mode(void)
         // 如果是二值化模式，使用大津法动态计算阈值
         if (display_mode == 1)
         {
-            threshold = otsu_get_threshold(mt9v03x_image[0], MT9V03X_W, MT9V03X_H);
+            threshold = (uint8)otsu_get_threshold(mt9v03x_image[0], MT9V03X_W, MT9V03X_H);
             image_threshold = threshold;
         }
